+++ src/strategies/mod.rs
@@ -28,6 +28,7 @@ pub struct Strategy {
 }
 
 impl CandleProcessor for Strategy {
+    type ReturnType = ();
     type ErrorType = StrategyError;
 
     /// Process all historical data
@@ -36,7 +37,7 @@ impl CandleProcessor for Strategy {
     /// data in the indicators is overwritten.
     ///
     /// This method is used upon initial load, or during backtesting.
-    fn process_candles(&mut self, candles: &DataFrame) -> Result<(), Self::ErrorType> {
+    fn process_candles(&mut self, candles: &DataFrame) -> Result<Self::ReturnType, Self::ErrorType> {
         for indicator in self.indicators.iter_mut() {
             indicator.process_candles(candles)
                 .map_err(|x| StrategyError::IndicatorError(x))?;